requirements.txt
click-alias==0.1.1a3
click-didyoumean==0.0.3
click-help-colors==0.4
click-spinner==0.1.8
click==6.7
delegator.py==0.1.0
emoji==0.5.0
mixpanel==4.3.2
prompt-toolkit==2.0.3
Pygments==2.2.0
raven==6.9.0
requests==2.19.1
storyscript==0.1.6
